% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secondary_structure.R
\name{features}
\alias{features}
\title{Get features from a secondary_structure object}
\usage{
features(obj)
}
\arguments{
\item{obj}{A secondary_structure object.}
}
\value{
A data.frame.
}
\description{
Get features from a secondary_structure object
}
\examples{
my_structure <- secondary_structure("a1", data.frame())
my_features <- features(my_structure)
}
