% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.ss.R
\name{filter_ss}
\alias{filter_ss}
\title{Filter a secondary structure based on a predicate}
\usage{
filter_ss(ss, pred)
}
\arguments{
\item{ss}{The secondary_structure object to filter.}

\item{pred}{The predicate, in dplyr format.}
}
\value{
A secondary_structure object with the same ID and filtered features.
}
\description{
Filter a secondary structure based on a predicate
}
\examples{
my_structure <- secondary_structure("a1", data.frame(Type = rep("a", 10),
Length = 1:10, Start = 1:10, Stop = 1:10))
# filter on a simple condition
structure_filtered <- filter_ss(my_structure, Length > 100)
# filter on a complex condition
structure_filtered_2 <- filter_ss(my_structure, (Start < 200 | Start > 500) & Length > 20)
}
