% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secondary_structure.R
\name{id}
\alias{id}
\title{Get id from a secondary_structure object}
\usage{
id(obj)
}
\arguments{
\item{obj}{A secondary_structure object.}
}
\value{
A character string.
}
\description{
Get id from a secondary_structure object
}
\examples{
my_structure <- secondary_structure("a1", data.frame())
my_id <- id(my_structure)
}
