% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secondary_structure.R
\name{is.secondary_structure}
\alias{is.secondary_structure}
\title{Check if an object is a secondary_structure}
\usage{
is.secondary_structure(obj)
}
\arguments{
\item{obj}{The object to be checked.}
}
\value{
Boolean.
}
\description{
Check if an object is a secondary_structure
}
\examples{
my_structure <- secondary_structure("a1", data.frame())
is.secondary_structure(my_structure)
}
