% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ss.R
\name{plot.secondary_structure}
\alias{plot.secondary_structure}
\title{NOT IMPLEMENTED}
\usage{
\method{plot}{secondary_structure}(x, ...)
}
\arguments{
\item{x}{A secondary_structure object.}

\item{...}{Other arguments, unused.}
}
\value{
Nothing.
}
\description{
This function is not implemented. Use ss_plot_one or ss_plot_many instead.
}
