% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secondary_structure.R
\name{print.secondary_structure}
\alias{print.secondary_structure}
\title{Print a secondary_structure object's basic information}
\usage{
\method{print}{secondary_structure}(x, ...)
}
\arguments{
\item{x}{A secondary_structure object.}

\item{...}{Other arguments for print(), not used but included for consistency.}
}
\value{
Nothing.
}
\description{
Print a secondary_structure object's basic information
}
\examples{
# call the generic instead of the method
my_structure <- secondary_structure("a1", data.frame())
print(my_structure)
}
