% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.ss.R
\name{read.ss}
\alias{read.ss}
\title{Load a secondary structure object from a TSV file}
\usage{
read.ss(path, header = TRUE, id = NULL)
}
\arguments{
\item{path}{The path to the TSV file.}

\item{header}{Whether the first line in the file contains TSV headers.}

\item{id}{The ID that should be stored in the object. Optional, inferred from the filename if not given.}
}
\value{
A list of class "secondary_structure".
}
\description{
Load a secondary structure object from a TSV file
}
\examples{
\dontrun{
structure <- read.ss("structure.tsv")
structure2 <- read.ss("~/data/other.tsv", header = FALSE, id = "dog")
}
}
