% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secondary_structure.R
\name{secondary_structure}
\alias{secondary_structure}
\title{Create a secondary structure object}
\usage{
secondary_structure(id, features)
}
\arguments{
\item{id}{The ID of the structure, use in reports and plots}

\item{features}{A data.frame containing the secondary structure features,
with columns containing type, length, start and end positions}
}
\value{
A list of class "secondary_structure"
}
\description{
Create a secondary structure object
}
\examples{
my_features <- data.frame()
my_structure <- secondary_structure("species1", my_features)
}
