% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss.summary.R
\name{ss.summary}
\alias{ss.summary}
\title{Pretty-print information about a secondary structure}
\usage{
ss.summary(ss)
}
\arguments{
\item{ss}{A secondary_structure object.}
}
\value{
Nothing.
}
\description{
Pretty-print information about a secondary structure
}
\examples{
\dontrun{
ss.summary(my_structure)
}
}
