% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ss.R
\name{ss_plot_many}
\alias{ss_plot_many}
\title{Plot multiple secondary structures for comparison}
\usage{
ss_plot_many(ss_list)
}
\arguments{
\item{ss_list}{A list of secondary_structure objects.}
}
\value{
A ggplot object.
}
\description{
Plot multiple secondary structures for comparison
}
\examples{
\dontrun{
ss_plot_many(my_structures)
}

}
